Restructure member details components, break it apart
