+++ administration/client/src/pages/MemberDetails.tsx
@@ -1,38 +1,34 @@
 import {
   ArrowLeftIcon,
-  ChatBubbleLeftIcon,
   ClockIcon,
   EnvelopeIcon,
   PencilIcon,
   TrashIcon,
 } from '@heroicons/react/24/outline';
-import { useMutation } from '@tanstack/react-query';
 import { useState } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
-import { MemberForm } from '../components/MemberForm';
 import { Modal } from '../components/Modal';
+import {
+  ApplicationNotes,
+  AuditHistory,
+  EditMemberForm,
+  EmailModal,
+  InfoCards,
+  LoadingStates,
+} from '../features/MemberDetails/components';
 import { useAuditLog } from '../hooks/useAudit';
 import {
   useDeleteMember,
   useMember,
   useUpdateMember,
 } from '../hooks/useMembers';
-import { api } from '../lib/api';
 import { MemberFormData } from '../types';
-import {
-  getAccessLevelColor,
-  getAccessLevelName,
-} from '../utils/authorization';
 
 export function MemberDetails() {
   const { id } = useParams<{ id: string }>();
   const navigate = useNavigate();
   const [isEditModalOpen, setIsEditModalOpen] = useState(false);
   const [isEmailModalOpen, setIsEmailModalOpen] = useState(false);
-  const [noteContent, setNoteContent] = useState('');
-  const [selectedTemplate, setSelectedTemplate] = useState('');
-  const [emailSubject, setEmailSubject] = useState('');
-  const [emailBody, setEmailBody] = useState('');
 
   const memberId = parseInt(id || '0', 10);
   const { data: member, isLoading, error } = useMember(memberId, !!id);
@@ -43,43 +39,6 @@ export function MemberDetails() {
   } = useAuditLog('member', memberId, !!id);
   const updateMember = useUpdateMember();
   const deleteMember = useDeleteMember();
-
-  const emailTemplates = [
-    {
-      id: 'welcome',
-      name: 'Welcome Email',
-      subject: 'Welcome to Our Community!',
-      body: "Hi {{first_name}},\n\nWelcome to our community! We're excited to have you join us.\n\nLorem ipsum dolor sit amet, consectetur adipiscing elit. Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.\n\nBest regards,\nThe Team",
-    },
-    {
-      id: 'reminder',
-      name: 'Event Reminder',
-      subject: 'Upcoming Event Reminder',
-      body: 'Hello {{preferred_name || first_name}},\n\nThis is a friendly reminder about our upcoming event.\n\nLorem ipsum dolor sit amet, consectetur adipiscing elit. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris.\n\nSee you there!\nEvent Team',
-    },
-    {
-      id: 'follow_up',
-      name: 'Follow-up',
-      subject: 'Following up with you',
-      body: 'Dear {{first_name}},\n\nI wanted to follow up on our recent conversation.\n\nLorem ipsum dolor sit amet, consectetur adipiscing elit. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore.\n\nPlease let me know if you have any questions.\n\nBest,\nAdmin Team',
-    },
-  ];
-
-  const addNoteMutation = useMutation({
-    mutationFn: async ({ content }: { content: string }) => {
-      const response = await api.post(`/members/${memberId}/notes`, {
-        content,
-        tags: [],
-      });
-      return response.data;
-    },
-    onSuccess: () => {
-      // Refetch audit log to show the new note
-      refetchAuditLog();
-      // Reset form
-      setNoteContent('');
-    },
-  });
 
   const handleUpdateMember = async (formData: MemberFormData) => {
     try {
@@ -114,22 +73,8 @@ export function MemberDetails() {
     }
   };
 
-  const handleTemplateSelect = (templateId: string) => {
-    setSelectedTemplate(templateId);
-    if (templateId) {
-      const template = emailTemplates.find((t) => t.id === templateId);
-      if (template) {
-        setEmailSubject(template.subject);
-        setEmailBody(template.body);
-      }
-    } else {
-      setEmailSubject('');
-      setEmailBody('');
-    }
-  };
-
-  const handleSendEmail = async () => {
-    if (!member || !emailSubject.trim() || !emailBody.trim()) {
+  const handleSendEmail = async (subject: string, body: string) => {
+    if (!member || !subject.trim() || !body.trim()) {
       alert('Please fill in both subject and body fields.');
       return;
     }
@@ -137,118 +82,24 @@ export function MemberDetails() {
     // In a real implementation, this would call an API to send the email
     // For now, we'll just show an alert
     alert(
-      `Email would be sent to ${member.email}\n\nSubject: ${emailSubject}\n\nBody: ${emailBody.substring(0, 100)}...\n\nThis is a placeholder implementation.`
+      `Email would be sent to ${member.email}\n\nSubject: ${subject}\n\nBody: ${body.substring(0, 100)}...\n\nThis is a placeholder implementation.`
     );
     setIsEmailModalOpen(false);
-    setSelectedTemplate('');
-    setEmailSubject('');
-    setEmailBody('');
   };
 
-  const handleEmailModalClose = () => {
-    setIsEmailModalOpen(false);
-    setSelectedTemplate('');
-    setEmailSubject('');
-    setEmailBody('');
+  const handleNoteAdded = () => {
+    refetchAuditLog();
   };
 
-  const handleSubmitNote = async (e: React.FormEvent) => {
-    e.preventDefault();
-
-    if (!noteContent.trim()) {
-      return;
-    }
-
-    try {
-      await addNoteMutation.mutateAsync({ content: noteContent.trim() });
-    } catch (error) {
-      console.error('Failed to add note:', error);
-      alert('Failed to add note. Please try again.');
-    }
-  };
-
-  const getDisplayName = (member: any) => {
-    return member.preferred_name
-      ? `${member.preferred_name} (${member.first_name}) ${member.last_name}`
-      : `${member.first_name} ${member.last_name}`;
-  };
-
-  const getStatusBadge = (member: any) => {
-    const isActive = member.flags & 1;
-    const isProfessional = member.flags & 2;
-
-    return (
-      <div className="flex flex-wrap gap-1">
-        <span
-          className={`inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium ${
-            isActive ? 'bg-green-100 text-green-800' : 'bg-red-100 text-red-800'
-          }`}
-        >
-          {isActive ? 'Active' : 'Inactive'}
-        </span>
-        {isProfessional && (
-          <span className="inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium bg-blue-100 text-blue-800">
-            Professional
-          </span>
-        )}
-        <span
-          className={`inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium ${getAccessLevelColor(
-            member.access_level || 1
-          )}`}
-        >
-          {getAccessLevelName(member.access_level || 1)}
-        </span>
-      </div>
-    );
-  };
-
-  if (isLoading) {
-    return (
-      <div className="flex items-center justify-center min-h-96">
-        <div className="text-gray-500">Loading member details...</div>
-      </div>
-    );
+  // Handle loading and error states
+  if (isLoading || error || !member) {
+    return <LoadingStates isLoading={isLoading} error={error} />;
   }
 
-  if (error || !member) {
-    // Check if it's a 403 Unauthorized error
-    const is403Error =
-      error &&
-      typeof error === 'object' &&
-      'response' in error &&
-      (error as any).response?.status === 403;
-
-    return (
-      <div className="text-center py-12">
-        <div className="text-red-600 mb-4">
-          {is403Error
-            ? "You don't have permission to access this member's details"
-            : 'Failed to load member details'}
-        </div>
-        <div className="space-x-4">
-          <button
-            onClick={() => navigate(is403Error ? '/dashboard' : '/members')}
-            className="bg-blue-600 text-white px-4 py-2 rounded-md hover:bg-blue-700"
-            data-testid="back-to-members-btn"
-          >
-            {is403Error ? 'Go to Dashboard' : 'Back to Members'}
-          </button>
-          {!is403Error && (
-            <button
-              onClick={() => window.location.reload()}
-              className="bg-gray-600 text-white px-4 py-2 rounded-md hover:bg-gray-700"
-            >
-              Retry
-            </button>
-          )}
-        </div>
-      </div>
-    );
-  }
+  const displayName = `${member.first_name} ${member.preferred_name ? `(${member.preferred_name}) ` : ''}${member.last_name}`;
 
   return (
     <div className="max-w-4xl mx-auto">
-      {/* Header */}
       <div className="flex items-center justify-between mb-6">
         <div className="flex items-center space-x-4">
           <button
@@ -263,7 +114,7 @@ export function MemberDetails() {
               className="text-3xl font-bold text-gray-900"
               data-testid="member-name-heading"
             >
-              {getDisplayName(member)}
+              {displayName}
             </h1>
             <p className="text-gray-600">{member.email}</p>
           </div>
@@ -296,224 +147,19 @@ export function MemberDetails() {
         </div>
       </div>
 
-      {/* Member Details */}
-      <div className="grid grid-cols-1 lg:grid-cols-2 gap-6">
-        {/* Basic Information */}
-        <div className="bg-white shadow rounded-lg p-6">
-          <h2 className="text-lg font-semibold text-gray-900 mb-4">
-            Basic Information
-          </h2>
-          <dl className="space-y-3">
-            <div>
-              <dt className="text-sm font-medium text-gray-500">Full Name</dt>
-              <dd className="text-sm text-gray-900">
-                {member.first_name} {member.last_name}
-              </dd>
-            </div>
-            {member.preferred_name && (
-              <div>
-                <dt className="text-sm font-medium text-gray-500">
-                  Preferred Name
-                </dt>
-                <dd className="text-sm text-gray-900">
-                  {member.preferred_name}
-                </dd>
-              </div>
-            )}
-            <div>
-              <dt className="text-sm font-medium text-gray-500">Email</dt>
-              <dd className="text-sm text-gray-900">{member.email}</dd>
-            </div>
-            {member.pronouns && (
-              <div>
-                <dt className="text-sm font-medium text-gray-500">Pronouns</dt>
-                <dd className="text-sm text-gray-900">{member.pronouns}</dd>
-              </div>
-            )}
-            <div>
-              <dt className="text-sm font-medium text-gray-500">Status</dt>
-              <dd className="text-sm text-gray-900">
-                {getStatusBadge(member)}
-              </dd>
-            </div>
-            <div>
-              <dt className="text-sm font-medium text-gray-500">
-                Member Since
-              </dt>
-              <dd className="text-sm text-gray-900">
-                {new Date(member.created_at).toLocaleDateString()}
-              </dd>
-            </div>
-          </dl>
-        </div>
+      <InfoCards member={member} />
 
-        {/* Additional Information */}
-        <div className="bg-white shadow rounded-lg p-6">
-          <h2 className="text-lg font-semibold text-gray-900 mb-4">
-            Additional Information
-          </h2>
-          <dl className="space-y-3">
-            {member.sponsor_notes && (
-              <div>
-                <dt className="text-sm font-medium text-gray-500">
-                  Sponsor Notes
-                </dt>
-                <dd className="text-sm text-gray-900 whitespace-pre-wrap">
-                  {member.sponsor_notes}
-                </dd>
-              </div>
-            )}
-            <div>
-              <dt className="text-sm font-medium text-gray-500">
-                Last Updated
-              </dt>
-              <dd className="text-sm text-gray-900">
-                {new Date(member.updated_at).toLocaleDateString()}
-              </dd>
-            </div>
-          </dl>
-        </div>
-      </div>
+      <ApplicationNotes memberId={memberId} onNoteAdded={handleNoteAdded} />
 
-      {/* Notes Section */}
-      <div className="mt-6 bg-white shadow rounded-lg p-6">
-        <h2 className="text-lg font-semibold text-gray-900 mb-3 flex items-center">
-          <ChatBubbleLeftIcon className="h-5 w-5 mr-2" />
-          Notes
-        </h2>
-
-        <form onSubmit={handleSubmitNote} className="mb-4">
-          <div className="flex gap-3">
-            <textarea
-              id="noteContent"
-              rows={2}
-              value={noteContent}
-              onChange={(e) => setNoteContent(e.target.value)}
-              className="flex-1 px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-blue-500 focus:border-blue-500 text-sm"
-              placeholder="Add a note about this member..."
-              data-testid="member-note-textarea"
-              disabled={addNoteMutation.isPending}
-            />
-            <button
-              type="submit"
-              disabled={!noteContent.trim() || addNoteMutation.isPending}
-              className="px-4 py-2 bg-purple-600 text-white rounded-md hover:bg-purple-700 disabled:opacity-50 disabled:cursor-not-allowed text-sm whitespace-nowrap self-start"
-            >
-              {addNoteMutation.isPending ? 'Adding...' : 'Add Note'}
-            </button>
-          </div>
-        </form>
-      </div>
-
-      {/* Activity History */}
       <div className="mt-6 bg-white shadow rounded-lg p-6">
         <h2 className="text-lg font-semibold text-gray-900 mb-4 flex items-center">
           <ClockIcon className="h-5 w-5 mr-2" />
-          Activity History
+          Audit History
         </h2>
-
         {auditLoading ? (
-          <div className="text-gray-500 text-sm">
-            Loading activity history...
-          </div>
-        ) : !auditLog || auditLog.length === 0 ? (
-          <div className="text-gray-500 text-sm">No activity recorded yet.</div>
+          <div className="text-gray-500 text-sm">Loading audit history...</div>
         ) : (
-          <div className="space-y-4">
-            {auditLog
-              .filter((entry) => entry.action !== 'view')
-              .map((entry) => (
-                <div
-                  key={entry.id}
-                  className="border-l-4 border-blue-200 pl-4 py-2"
-                >
-                  <div className="flex items-center justify-between">
-                    <div className="flex items-center space-x-2">
-                      <span
-                        className={`inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium ${
-                          entry.action === 'create'
-                            ? 'bg-green-100 text-green-800'
-                            : entry.action === 'update'
-                              ? 'bg-yellow-100 text-yellow-800'
-                              : entry.action === 'delete'
-                                ? 'bg-red-100 text-red-800'
-                                : entry.action === 'view'
-                                  ? 'bg-blue-100 text-blue-800'
-                                  : entry.action === 'note'
-                                    ? 'bg-purple-100 text-purple-800'
-                                    : 'bg-gray-100 text-gray-800'
-                        }`}
-                      >
-                        {entry.action.charAt(0).toUpperCase() +
-                          entry.action.slice(1)}
-                      </span>
-                      <span className="text-sm text-gray-600">
-                        {new Date(entry.created_at).toLocaleString()}
-                      </span>
-                    </div>
-                    <span className="text-xs text-gray-400">
-                      Session:{' '}
-                      {(entry.user_session_id || 'unknown').slice(0, 8)}... |
-                      IP: {entry.user_ip}
-                    </span>
-                  </div>
-
-                  {entry.action === 'update' &&
-                    entry.oldValues &&
-                    entry.newValues && (
-                      <div className="mt-2 text-sm">
-                        <div className="text-gray-600 mb-1">Changes made:</div>
-                        <div className="bg-gray-50 rounded p-2 space-y-1">
-                          {Object.keys(entry.newValues).map((key) => {
-                            const oldValue = entry.oldValues?.[key];
-                            const newValue = entry.newValues?.[key];
-
-                            if (
-                              oldValue !== newValue &&
-                              key !== 'id' &&
-                              key !== 'updated_at'
-                            ) {
-                              return (
-                                <div key={key} className="text-xs">
-                                  <span className="font-medium capitalize">
-                                    {key.replace(/_/g, ' ')}:
-                                  </span>
-                                  <span className="text-red-600 line-through ml-1">
-                                    {String(oldValue || '(empty)')}
-                                  </span>
-                                  <span className="mx-1">→</span>
-                                  <span className="text-green-600">
-                                    {String(newValue || '(empty)')}
-                                  </span>
-                                </div>
-                              );
-                            }
-                            return null;
-                          })}
-                        </div>
-                      </div>
-                    )}
-
-                  {entry.action === 'create' && entry.newValues && (
-                    <div className="mt-2 text-sm">
-                      <div className="text-gray-600">
-                        Member created with initial data
-                      </div>
-                    </div>
-                  )}
-
-                  {entry.action === 'note' && entry.metadata && (
-                    <div className="mt-2 text-sm">
-                      <div className="bg-purple-50 rounded p-3 border-l-2 border-purple-200">
-                        <div className="text-gray-800 whitespace-pre-wrap">
-                          {entry.metadata.content}
-                        </div>
-                      </div>
-                    </div>
-                  )}
-                </div>
-              ))}
-          </div>
+          <AuditHistory auditLog={auditLog} />
         )}
       </div>
 
@@ -523,7 +169,7 @@ export function MemberDetails() {
         onClose={() => setIsEditModalOpen(false)}
         title="Edit Member"
       >
-        <MemberForm
+        <EditMemberForm
           member={member}
           onSubmit={handleUpdateMember}
           onCancel={() => setIsEditModalOpen(false)}
@@ -531,101 +177,12 @@ export function MemberDetails() {
         />
       </Modal>
 
-      {/* Email Modal */}
-      <Modal
+      <EmailModal
         isOpen={isEmailModalOpen}
-        onClose={handleEmailModalClose}
-        title="Send Email"
-      >
-        <div className="space-y-4">
-          <div>
-            <p className="text-sm text-gray-600 mb-3">
-              Send email to:{' '}
-              <span className="font-medium">{member?.email}</span>
-            </p>
-            <label
-              htmlFor="emailTemplate"
-              className="block text-sm font-medium text-gray-700 mb-2"
-            >
-              Choose Template (Optional)
-            </label>
-            <select
-              id="emailTemplate"
-              value={selectedTemplate}
-              onChange={(e) => handleTemplateSelect(e.target.value)}
-              className="w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-blue-500 focus:border-blue-500"
-            >
-              <option value="">Write your own email...</option>
-              {emailTemplates.map((template) => (
-                <option key={template.id} value={template.id}>
-                  {template.name}
-                </option>
-              ))}
-            </select>
-          </div>
-
-          <div>
-            <label
-              htmlFor="emailSubject"
-              className="block text-sm font-medium text-gray-700 mb-2"
-            >
-              Subject
-            </label>
-            <input
-              type="text"
-              id="emailSubject"
-              value={emailSubject}
-              onChange={(e) => setEmailSubject(e.target.value)}
-              className="w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-blue-500 focus:border-blue-500"
-              placeholder="Enter email subject..."
-            />
-          </div>
-
-          <div>
-            <label
-              htmlFor="emailBody"
-              className="block text-sm font-medium text-gray-700 mb-2"
-            >
-              Message
-            </label>
-            <textarea
-              id="emailBody"
-              rows={8}
-              value={emailBody}
-              onChange={(e) => setEmailBody(e.target.value)}
-              className="w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-blue-500 focus:border-blue-500"
-              placeholder="Write your email message here..."
-            />
-          </div>
-
-          <div className="bg-blue-50 p-3 rounded-md">
-            <h4 className="font-medium text-blue-900 mb-1">
-              Available Variables:
-            </h4>
-            <p className="text-sm text-blue-700">
-              You can use these placeholders: <code>{'{{first_name}}'}</code>,{' '}
-              <code>{'{{last_name}}'}</code>,{' '}
-              <code>{'{{preferred_name}}'}</code>, <code>{'{{email}}'}</code>
-            </p>
-          </div>
-
-          <div className="flex justify-end space-x-3">
-            <button
-              onClick={handleEmailModalClose}
-              className="px-4 py-2 text-gray-700 bg-gray-200 rounded-md hover:bg-gray-300"
-            >
-              Cancel
-            </button>
-            <button
-              onClick={handleSendEmail}
-              disabled={!emailSubject.trim() || !emailBody.trim()}
-              className="px-4 py-2 bg-green-600 text-white rounded-md hover:bg-green-700 disabled:opacity-50 disabled:cursor-not-allowed"
-            >
-              Send Email
-            </button>
-          </div>
-        </div>
-      </Modal>
+        onClose={() => setIsEmailModalOpen(false)}
+        member={member}
+        onSendEmail={handleSendEmail}
+      />
     </div>
   );
 }
